+++ src/test/java/net/minecraft/class_2960.java
@@ -1,0 +1,5 @@
+package net.minecraft;
+
+public record class_2960(String value) {
+    public static class_2960 method_60654(String value) { return new class_2960(value); }
+}